+++ NEWS.md
@@ -19,7 +19,8 @@
 
 ## New features
 
-* Made "newperson" optional in `prep_data()` (#73)
+* Added a default `print.sr_model()` method for `sr_model` objects (#114)
+* Made `newperson` optional in `prep_data()` (#73)
 * Including fitted and residual values as data frame in run_mod output. (#101)
 * Added  `plot_predicted_curve()` with support for faceting by multiple IDs (#68)
 * Replacing old data object with new run_mod output (#102)